[#155522818] Feature: Set aside cases in the case table
Case table submenu with actions "set aside selected", "set aside
unselected", and "restore set aside cases"

+++ apps/dg/controllers/data_context.js
@@ -1017,6 +1017,7 @@ DG.DataContext = SC.Object.extend((function() // closure
    */
   deleteCasesAndChildren: function(iChange) {
     var deletedCases = [];
+    var tSetAside = iChange.setAside;
 
     iChange.ids = [];
     iChange.collectionIDs = {};
@@ -1040,7 +1041,12 @@ DG.DataContext = SC.Object.extend((function() // closure
       // keep track of the affected collections
       iChange.collectionIDs[tCollection.get('id')] = tCollection;
 
-      tCollection.deleteCase(iCase);
+      if (tSetAside) {
+        tCollection.setAsideCase(iCase);
+      } else {
+        tCollection.deleteCase(iCase);
+      }
+
       if (tParent) {
         if (tParent.children.length === 0) {
           doDelete(tParent);
@@ -2557,7 +2563,15 @@ DG.DataContext = SC.Object.extend((function() // closure
         this.addCollection( collection);
       }.bind(this));
     }
-  }
+  },
+
+    restoreSetAsideCases: function () {
+      var items = this.get('dataSet').restoreSetAsideItems();
+      var count = items.length;
+      var results = this.regenerateCollectionCases(null, 'createCases');
+      DG.log("Restored " + count + " items in context \"" + this.get('name') + "\"");
+      DG.log("Restored %@ cases".loc(results.createdCases && results.createdCases.length));
+    }
 
   }; // end return from closure
